+++ services/location-service/internal/models/models.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Location struct {
-	ID          int    `db:"id" json:"id"`
+	ID          int32  `db:"id" json:"id"`
 	Name        string `db:"name" json:"name"`
 	Coordinates string `db:"coordinates" json:"coordinates"`
 }
@@ -14,7 +14,7 @@ func ToProto(locations []*Location) []*locationPb.Location {
 	locationsProto := make([]*locationPb.Location, len(locations))
 	for idx, location := range locations {
 		locationsProto[idx] = &locationPb.Location{
-			LocationId:  int32(location.ID),
+			LocationId:  location.ID,
 			Name:        location.Name,
 			Coordinates: location.Coordinates,
 		}
